优化并发率控制，详情页封面图背景色为主题色

+++ app/src/main/java/io/legado/app/exception/ConcurrentException.kt
@@ -5,4 +5,4 @@ package io.legado.app.exception
 /**
  * 并发限制
  */
-class ConcurrentException(msg: String, val waitTime: Int) : NoStackTraceException(msg)
+class ConcurrentException(msg: String, val waitTime: Long) : NoStackTraceException(msg)
